Fully modernized the registration package.
  -- Rewrote register.py with RuntimeContext architecture replacing the legacy ops dictionary pattern.
  -- Renamed registration_wrapper to register_plane and save_registration_outputs_to_ops was removed.
  -- Added _ReferenceData and _BatchRegistrationResult dataclasses for structured data handling.
  -- Added RegistrationBlocks type alias for complex tuple return types.
  -- Replaced all magic numbers with module-level constants across deformation.py, diffeomorphic.py, pyramid.py, and register.py.
  -- Added type: ignore[import-untyped] comments to numba imports in deformation.py, nonrigid.py, and spline_grid.py.
  -- Fixed type narrowing in diffeomorphic.py using isinstance checks for Deformation and ndarray cache entries.
  -- Added None validation for _pyramids attribute with RuntimeError in diffeomorphic.py.
  -- Changed compute_pc_metrics return type from RuntimeContext to None for in-place modification semantics.
  -- Updated bidiphase_correction.py type annotations from np.int16 to np.float32.
  -- Simplified nested conditional in spline_grid.py to single return statement.
  -- Removed unused type ignore comment from utils.py vectorize decorator.
  -- Deleted zalign.py module as z-position computation is no longer part of the registration pipeline.
  -- Updated __init__.py exports to reflect removed and renamed functions.
  -- Converted all function calls to use keyword arguments for clarity.
  -- Ensured all docstrings follow Google style with proper Args, Returns, and Raises sections.

+++ src/sl_suite2p/registration/bidiphase_correction.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
     from numpy.typing import NDArray
 
 
-def compute_bidirectional_phase_offset(frames: NDArray[np.int16]) -> int:
+def compute_bidirectional_phase_offset(frames: NDArray[np.float32]) -> int:
     """Computes the bidirectional phase offset from a sample of imaging frames.
 
     Bidirectional scanning microscopes acquire alternating lines in opposite directions, which can introduce a
@@ -53,7 +53,7 @@ def compute_bidirectional_phase_offset(frames: NDArray[np.int16]) -> int:
 
 
 def apply_bidirectional_phase_correction(
-    frames: NDArray[np.int16],
+    frames: NDArray[np.float32],
     bidirectional_phase_offset: int,
 ) -> None:
     """Applies bidirectional phase correction to imaging frames in-place.
